work on validation builder UI. move eido (#94)
